Extract build and base command into own files; Use common pattern for .go files with commands
